Now lists files with the extension ".tar.gz" and allows the usr to select them.
Thanks to openAI for this one! I was too tired to figure it out after hours of programming, so I just defaulted to the old method.

Comments are properly included, and now it works as intended. It lists all files ending with the extension of ".tar.gz", allows the user to select them, and then stores what the user selected into a variable called "output_file". Then, we iterate with that information to easily select specific DBs to restore.